There are possible problems when jiffies (ticks) which still are int
wrap around.  Also given this did not touch every single place some
checks may be broken now.

Reported by:	markj
This reverts commit fd27f86dd71b7ff1df6981297095b88d1d29652e.

+++ sys/compat/linuxkpi/common/src/linux_compat.c
@@ -1934,7 +1934,7 @@ linux_timer_callback_wrapper(void *context)
 }
 
 int
-mod_timer(struct timer_list *timer, unsigned long expires)
+mod_timer(struct timer_list *timer, int expires)
 {
 	int ret;
